+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-10 02:53:59
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: eab9217db44872762d63bfa99e58a43ebe01062b
+
+## [ForwardWidgets-2kuai] - 2025-10-10 02:53:59
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: da4ac4e07d8c0f437d4e482534ccc5d04e4da7a4
+
+## [ForwardWidgets-pack1r] - 2025-10-10 02:53:59
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: c93e40573035cd04fd48a2b555c579be54202916
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-09 02:53:18
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main